feat: fix Google Sign-In, App Check debug, and optimize location management
- Added Firebase App Check Debug dependency and provider
- Registered SHA fingerprints in Firebase
- Fixed crash and visual borders in HomeFragment
- Implemented intelligent search in ManageLocations
- Optimized locations loading with parallel coroutines
- Standardized XP types to Long across multiple domain services

+++ app/src/main/java/com/futebadosparcas/domain/ranking/PostGameEventEmitter.kt
@@ -33,7 +33,7 @@ class PostGameEventEmitter @Inject constructor() {
     fun createSummary(
         gameId: String,
         result: PlayerProcessingResult,
-        previousXp: Int,
+        previousXp: Long,
         previousLevel: Int
     ): PostGameSummary {
         return PostGameSummary(
